rename models and clean configs
