+++ conceptarium/conf/loss/composite.yaml
@@ -1,14 +1,8 @@
 # =============================================================
-# Composite loss settings
-#
-# Modular loss composition: define a list of loss terms.
-# The learner wraps them internally into a weighted sum.
-# Each term that needs 'annotations' gets it injected at
-# runtime by run_experiment.py.
-#
-# Set loss_weights at the top level of the config (e.g. in
-# sweep.yaml) to control per-term weighting:
-#   loss_weights: [1.0, 0.5]
+# A list of loss terms, summed with weight 1.0 each by the
+# learner. Terms needing `annotations` get it injected by
+# run_experiment.py. For per-term weights, use a CompositeLoss
+# with its own `weights:` instead (see conf/loss/cbvae.yaml).
 # =============================================================
 - _target_: "torch_concepts.nn.WeightedConceptLoss"
   concept_weight: 10 # weight applied to concepts
